auto-google-search/requirements.txt
python-dotenv
requests
pytest
ruff
click
google-custom-search
pyyaml
